refactor of static optimization
I removed qddot as entry points since we may decide to actually target qddot and it would not be the same cost function.

Removed bunch of redundant members of s2mStaticOptimization and compressed them into the vectored ones.

Removed the storage of final solution, since storing the actual TNLP stores more information. An interface to get the final results is still to be written though. 

Reinserted useResidualTorque which had been removed for some reason?

If no initial guess is provided, a default one is sent. If one initial guess is provided for the multiple time optimization, this one is used for the first frame then the previous solution is used as the next first guess